refactor(#45): Studies data class 패키지 이동
