+++ app/src/main/java/com/ssafy/neegongnaegong/data/remote/StudiesApi.kt
@@ -1,6 +1,6 @@
 package com.ssafy.neegongnaegong.data.remote
 
-import com.ssafy.neegongnaegong.domain.model.Studies
+import com.ssafy.neegongnaegong.domain.model.studies.Studies
 import retrofit2.http.GET
 
 interface StudiesApi {
