+++ app/src/main/java/com/ssafy/neegongnaegong/domain/model/studies/Studies.kt
@@ -1,4 +1,4 @@
-package com.ssafy.neegongnaegong.domain.model
+package com.ssafy.neegongnaegong.domain.model.studies
 
 data class Studies(
     val id: Long,
