+++ app/src/main/java/com/ssafy/neegongnaegong/domain/repository/StudiesRepository.kt
@@ -1,6 +1,6 @@
 package com.ssafy.neegongnaegong.domain.repository
 
-import com.ssafy.neegongnaegong.domain.model.Studies
+import com.ssafy.neegongnaegong.domain.model.studies.Studies
 
 interface StudiesRepository {
     suspend fun getStudies(): List<Studies>
