+++ app/src/main/java/com/ssafy/neegongnaegong/domain/usecase/GetStudiesUseCase.kt
@@ -1,6 +1,6 @@
 package com.ssafy.neegongnaegong.domain.usecase
 
-import com.ssafy.neegongnaegong.domain.model.Studies
+import com.ssafy.neegongnaegong.domain.model.studies.Studies
 import com.ssafy.neegongnaegong.domain.repository.StudiesRepository
 import javax.inject.Inject
