+++ app/src/main/java/com/ssafy/neegongnaegong/presentation/group/component/drawer/StudiesDrawer.kt
@@ -220,8 +220,7 @@ fun CircleIcon(
                 .background(
                     color = MaterialTheme.colorScheme.surfaceVariant,
                     shape = CircleShape,
-                )
-                .clickable {
+                ).clickable {
                     onClick()
                 },
         contentAlignment = Alignment.Center,
